server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# YouTube Download
yt-dlp>=2024.8.6

# Media Processing
ffmpeg-python==0.2.0

# AI/ML
numpy==1.24.3
openai-whisper==20231117
openai==1.3.7
google-generativeai==0.3.1
torch==2.1.1
torchaudio==2.1.1

# Utilities
httpx==0.25.2
aiofiles==23.2.1
